Refactor signing
Instead of hard coding a signing key id in the script, you now
specify --signing-key=<key file> on the command line if you want
to sign the packages.  The key file has to be appropriate for
the package type being created.  For "deb", it's a file containing
your signing key id.  For "rpm" it could be your rsa or gpg key.