Squash a few mote packaging bugs
The spec file for duetruntime was excluding some files when in
Release mode that it shouldn't have been.

The signing key wasn't working when a filename was supplied with
a "~" at the front.  To help tell the difference between a key id
and a key file, key file names must begin with '@'.

For instance, both of these are now valid:
A plain key id:
--signing-key=7F7F6E6E
A file containing a key id:
--signing-key=@~/.keys/dsf-signing-key
--signing-key=@/home/user/.keys/dsf-signing-key

Also FYI...  The key id only needs to be long enough to uniquely
identify the fingerprint of your GPG signing key.  8 characters
is plenty.  In fact, it doesn't even have to be the key's
fingerprint.  You could use the email address specified when
you created the key as long as it's unique.
--signing-key="<me@somewhere.com>"
Anything that GPG accepts as a key search item can be used.