Update build instructions to specify .Net Core 3.1
The instructions specified to "get the latest .net sdk", but that's a bit confusing when there are 3 distinct branches of .net: core 2.1, core 3.1 and 5.0.  Making it more confusing, MS doesn't use the word "core" in 5.0.  A person might think "oh, 5.0 must be the latest", but it appears that DSF uses the 3.1 branch, the latest version of which is currently 3.1.11 (at the time of this comment.)  This change just clarifies "CORE 3.1"