Added way to flag when plugins have fully started
Plugins can now define in their manifest if they notify DSF when they
have fully started. This is be achieved using the new
NotifyPluginStarted command in combination with the new sbcNotifyStarted
property in the plugin manifest. That way, DSF can wait for plugins to
finish their start-up sequence before dsf-config.g is started. As a
consequence, a G4 delay in dsf-config.g should not be needed any more